errorutil: refactoring
