+++ cmd/errorutil/internal/config/app.go
@@ -1,0 +1,6 @@
+package config
+
+const (
+	App     = "errorutil"
+	LogFile = App + ".log"
+)
