+++ cmd/errorutil/internal/config/logging.go
@@ -1,0 +1,13 @@
+package config
+
+import "github.com/sirupsen/logrus"
+
+func Logging(verbose bool) {
+	logrus.SetFormatter(&logrus.TextFormatter{})
+	//logrus.SetFormatter(&logrus.JSONFormatter{})
+	if verbose {
+		logrus.SetLevel(logrus.DebugLevel)
+	} else {
+		logrus.SetLevel(logrus.InfoLevel)
+	}
+}
